Working SPI RAM based FIFO

+++ include/fifo.h
@@ -1,13 +1,13 @@
 #ifndef INCLUDE_FIFO_H_
 #define INCLUDE_FIFO_H_
 
-#include <stdint.h>
+#include <stddef.h>
 
 int fifo_init(void);
-void fifo_enqueue(const void *data, uint32_t len);
-void fifo_dequeue(void *data, uint32_t len);
-uint32_t fifo_fill(void);
-uint32_t fifo_free(void);
-uint32_t fifo_size(void);
+void fifo_enqueue(const void *data, size_t len);
+void fifo_dequeue(void *data, size_t len);
+size_t fifo_fill(void);
+size_t fifo_free(void);
+size_t fifo_size(void);
 
 #endif /* INCLUDE_FIFO_H_ */
